Move to attribute sub-package
